All tests pass now.
The problem was with certain Functions being called
without a parameter, if that parameter is not given in the initial
Function definition. I think that's the correct behavior...

The problem occurred twice in the tests. In one case,
utils.createfourierdesignmatrix_red was being created
(in FourierBasisCommonGP) without the Tspan parameter.
That was easy to fix.

In the other case, a Parameter sampler method was being
called without the `size` parameter, since that is not
a standard parameter for the prior Function, and the sampler is
called as an "alternate func" using the prior instance
(maybe that was not the most enlightened choice).
At this time, I'm handling `size` as a special case in
Function.__call__, but this may come to bite us later.

I have also modified utils.createfourierdesignmatrix_red
so that partially overlapping mode sets created with
different nmodes and the same Tspan have perfectly
overlapping frequencies.